refactor: move tests to integration tests
