+++ crates/rattler-bin/Cargo.toml
@@ -32,7 +32,7 @@ rattler = { version = "0.5.0", path = "../rattler", default-features = false }
 rattler_networking = { version = "0.5.0", path = "../rattler_networking", default-features = false }
 rattler_conda_types = { version = "0.5.0", path = "../rattler_conda_types" }
 rattler_repodata_gateway = { version = "0.5.0", path = "../rattler_repodata_gateway", features = ["sparse"], default-features = false }
-rattler_solve = { version = "0.5.0", path = "../rattler_solve", features = ["libsolv-rs"] }
+rattler_solve = { version = "0.5.0", path = "../rattler_solve", features = ["libsolv_rs", "libsolv-sys"] }
 rattler_virtual_packages = { version = "0.5.0", path = "../rattler_virtual_packages" }
 reqwest = { version = "0.11.18", default-features = false }
 tokio = { version = "1.28.2", features = ["rt-multi-thread", "macros"] }
